+++ Stream_support/include/CGAL/IO/PLY/PLY_reader.h
@@ -11,6 +11,7 @@
 #ifndef CGAL_IO_PLY_PLY_READER_H
 #define CGAL_IO_PLY_PLY_READER_H
 
+#include <CGAL/IO/PLY.h>
 #include <CGAL/Container_helper.h>
 #include <CGAL/IO/io.h>
 #include <CGAL/type_traits/is_iterator.h>
