+++ Stream_support/include/CGAL/IO/PLY/read_ply_points.h
@@ -36,99 +36,7 @@ namespace CGAL {
 
 namespace IO {
 
-#ifdef DOXYGEN_RUNNING // Document some parts from Stream_support here for convenience
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   Class used to identify a %PLY property as a type and a name.
-
-   \sa `read_PLY_with_properties()`
-*/
-template <typename T>
-struct PLY_property
-{
-  typedef T type;
-  const char* name;
-  PLY_property(const char* name) : name(name) { }
-};
-
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   Generates a %PLY property handler to read 3D points. Points are
-   constructed from the input using 3 %PLY properties of type `FT`
-   and named `x`, `y` and `z`. `FT` is `float` if the points use
-   `CGAL::Simple_cartesian<float>` and `double` otherwise.
-
-   \tparam PointMap the property map used to store points.
-
-   \sa `read_PLY_with_properties()`
-   \sa \ref IOStreamPLY
-*/
-template <typename PointMap>
-std::tuple<PointMap,
-           typename Kernel_traits<typename PointMap::value_type>::Kernel::Construct_point_3,
-           PLY_property<FT>, PLY_property<FT>, PLY_property<FT> >
-make_ply_point_reader(PointMap point_map);
-
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   Generates a %PLY property handler to read 3D normal
-   vectors. Vectors are constructed from the input using 3 PLY
-   properties of type `FT` and named `nx`, `ny` and `nz`. `FT`
-   is `float` if the points use `CGAL::Simple_cartesian<float>` and
-   `double` otherwise.
-
-   \tparam VectorMap the property map used to store vectors.
-
-   \sa `read_PLY_with_properties()`
-   \sa \ref IOStreamPLY
-*/
-template <typename VectorMap>
-std::tuple<VectorMap,
-           typename Kernel_traits<typename VectorMap::value_type>::Kernel::Construct_vector_3,
-           PLY_property<FT>, PLY_property<FT>, PLY_property<FT> >
-make_ply_normal_reader(VectorMap normal_map);
-
-#endif // DOXYGEN_RUNNING
-
-/**
-  \ingroup PkgPointSetProcessing3IOPly
-
-  \brief reads user-selected points properties from a .ply stream (ASCII or binary).
-
-  Potential additional point properties and faces are ignored.
-
-  Properties are handled through a variadic list of property
-  handlers. A `PropertyHandler` can either be:
-
-  - A `std::pair<PropertyMap, PLY_property<T> >` if the user wants
-  to read a %PLY property as a scalar value T (for example, storing
-  an `int` %PLY property into an `int` variable).
-
-  - A `std::tuple<PropertyMap, Constructor,
-  PLY_property<T>...>` if the user wants to use one or several PLY
-  properties to construct a complex object (for example, storing 3
-  `uchar` %PLY properties into a %Color object that can for example
-  be a `std::array<unsigned char, 3>`). In that case, the
-  second element of the tuple should be a functor that constructs
-  the value type of `PropertyMap` from N objects of types `T`.
-
-  \attention To read a binary file, the flag `std::ios::binary` must be set during the creation of the `ifstream`.
-
-  \tparam OutputIteratorValueType type of objects that can be put in `PointOutputIterator`.
-  It must be a model of `DefaultConstructible` and defaults to `value_type_traits<PointOutputIterator>::%type`.
-  It can be omitted when the default is fine.
-  \tparam PointOutputIterator iterator over output points.
-  \tparam PropertyHandler handlers to recover properties.
-
-  \returns `true` if reading was successful, `false` otherwise.
-
-  \sa \ref IOStreamPLY
-  \sa `make_ply_point_reader()`
-  \sa `make_ply_normal_reader()`
-*/
+// documentation in ../PLY.h
 template <typename OutputIteratorValueType,
           typename PointOutputIterator,
           typename ... PropertyHandler>
@@ -261,63 +169,14 @@ bool read_PLY(std::istream& is,
                                   make_ply_normal_reader(normal_map));
 }
 
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   \brief reads points (positions + normals, if available), using the \ref IOStreamPLY.
-
-   Potential additional point properties and faces are ignored.
-
-   \tparam OutputIteratorValueType type of objects that can be put in `PointOutputIterator`.
-   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<PointOutputIterator>::%type`.
-   It can be omitted when the default is fine.
-   \tparam PointOutputIterator iterator over output points.
-   \tparam NamedParameters a sequence of \ref bgl_namedparameters "Named Parameters"
-
-   \param fname input file name.
-   \param output output iterator over points.
-   \param np optional sequence of \ref bgl_namedparameters "Named Parameters" among the ones listed below.
-
-   \cgalNamedParamsBegin
-     \cgalParamNBegin{use_binary_mode}
-       \cgalParamDescription{indicates whether data should be read in binary (`true`) or in \ascii (`false`)}
-       \cgalParamType{Boolean}
-       \cgalParamDefault{`true`}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{point_map}
-       \cgalParamDescription{a property map associating points to the elements of the point range}
-       \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Point_3`}
-       \cgalParamDefault{`CGAL::Identity_property_map<geom_traits::Point_3>`}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{normal_map}
-       \cgalParamDescription{a property map associating normals to the elements of the point range}
-       \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Vector_3`}
-       \cgalParamDefault{If this parameter is omitted, normals in the input stream are ignored.}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{geom_traits}
-       \cgalParamDescription{an instance of a geometric traits class}
-       \cgalParamType{a model of `Kernel`}
-       \cgalParamDefault{a \cgal Kernel deduced from the point type, using `CGAL::Kernel_traits`}
-     \cgalParamNEnd
-   \cgalNamedParamsEnd
-
-   \returns `true` if reading was successful, `false` otherwise.
-
-   \sa \ref IOStreamPLY
-   \sa `read_PLY_with_properties()`
-*/
+// documentation in ../PLY.h
 template <typename OutputIteratorValueType,
           typename PointOutputIterator,
-          typename CGAL_NP_TEMPLATE_PARAMETERS>
+          typename CGAL_NP_TEMPLATE_PARAMETERS_NO_DEFAULT>
 bool read_PLY(const std::string& fname,
               PointOutputIterator output,
-              const CGAL_NP_CLASS& np = parameters::default_values()
-#ifndef DOXYGEN_RUNNING
-              , std::enable_if_t<CGAL::is_iterator<PointOutputIterator>::value>* = nullptr
-#endif
+              const CGAL_NP_CLASS& np,
+              std::enable_if_t<CGAL::is_iterator<PointOutputIterator>::value>*
               )
 {
   const bool binary = CGAL::parameters::choose_parameter(CGAL::parameters::get_parameter(np, internal_np::use_binary_mode), true);
