+++ Stream_support/include/CGAL/IO/PLY/write_ply_points.h
@@ -38,76 +38,8 @@ namespace CGAL {
 
 namespace IO {
 
-#ifdef DOXYGEN_RUNNING // Document some parts from Stream_support here for convenience
-  /**
-     \ingroup PkgPointSetProcessing3IOPly
-
-     Generates a %PLY property handler to write 3D points. Points are
-     written as 3 %PLY properties of type `FT` and named `x`, `y` and
-     `z`. `FT` is `float` if the points use
-     `CGAL::Simple_cartesian<float>` and `double` otherwise.
-
-     \tparam PointMap the property map used to store points.
-
-     \sa `write_PLY_with_properties()`
-     \sa \ref IOStreamPLY
-  */
-  template <typename PointMap>
-  std::tuple<PointMap, PLY_property<FT>, PLY_property<FT>, PLY_property<FT> >
-  make_ply_point_writer(PointMap point_map);
-
-  /**
-     \ingroup PkgPointSetProcessing3IOPly
-
-     Generates a %PLY property handler to write 3D normal
-     vectors. Vectors are written as 3 %PLY properties of type `FT`
-     and named `nx`, `ny` and `nz`. `FT` is `float` if the vectors use
-     `CGAL::Simple_cartesian<float>` and `double` otherwise.
-
-     \tparam VectorMap the property map used to store vectors.
-
-     \sa `write_PLY_with_properties()`
-     \sa \ref IOStreamPLY
-  */
-  template <typename VectorMap>
-  std::tuple<VectorMap, PLY_property<FT>, PLY_property<FT>, PLY_property<FT> >
-  make_ply_normal_writer(VectorMap normal_map);
-#endif
-
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   \brief writes the range of `points` with properties using \ref IOStreamPLY.
-
-   Properties are handled through a variadic list of property
-   handlers. A `PropertyHandler` can either be:
-
-   - A `std::pair<PropertyMap, PLY_property<T> >` if the user wants
-   to write a scalar value T as a %PLY property (for example, writing
-   an `int` variable as an `int` %PLY property).
-
-   - A `std::tuple<PropertyMap, PLY_property<T>...>` if the
-   user wants to write a complex object as several %PLY
-   properties. In that case, a specialization of `Output_rep` must
-   be provided for `PropertyMap::value_type` that handles both ASCII
-   and binary output (see `CGAL::IO::get_mode()`).
-
-   \attention To write to a binary file, the flag `std::ios::binary` must be set during the creation
-              of the `ofstream`, and the \link PkgStreamSupportEnumRef `IO::Mode` \endlink
-              of the stream must be set to `BINARY`.
-
-   \tparam PointRange is a model of `ConstRange`. The value type of
-                      its iterator is the key type of the `PropertyMap` objects provided
-                      within the `PropertyHandler` parameter.
-   \tparam PropertyHandler handlers to recover properties.
-
-   \returns `true` if writing was successful, `false` otherwise.
-
-   \sa \ref IOStreamPLY
-   \sa `make_ply_point_writer()`
-   \sa `make_ply_normal_writer()`
-*/
-template <typename PointRange,
+  // documented in ../PLY.h
+  template <typename PointRange,
           typename ... PropertyHandler>
   bool write_PLY_with_properties(std::ostream& os, ///< output stream.
                                  const PointRange& points, ///< input point range.
@@ -138,61 +70,12 @@ template <typename PointRange,
   return !os.fail();
 }
 
-/**
-   \ingroup PkgPointSetProcessing3IOPly
-
-   \brief writes the range of `points` (positions + normals, if available) using \ref IOStreamPLY.
-
-   \attention To write to a binary file, the flag `std::ios::binary` must be set during the creation
-              of the `ofstream`, and the \link PkgStreamSupportEnumRef `IO::Mode` \endlink
-              of the stream must be set to `BINARY`.
-
-   \tparam PointRange is a model of `ConstRange`. The value type of
-                      its iterator is the key type of the named parameter `point_map`.
-   \tparam NamedParameters a sequence of \ref bgl_namedparameters "Named Parameters"
-
-   \param os output stream
-   \param points input point range
-   \param np an optional sequence of \ref bgl_namedparameters "Named Parameters" among the ones listed below
-
-   \cgalNamedParamsBegin
-     \cgalParamNBegin{point_map}
-       \cgalParamDescription{a property map associating points to the elements of the point range}
-       \cgalParamType{a model of `ReadablePropertyMap` with value type `geom_traits::Point_3`}
-       \cgalParamDefault{`CGAL::Identity_property_map<geom_traits::Point_3>`}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{normal_map}
-       \cgalParamDescription{a property map associating normals to the elements of the point range}
-       \cgalParamType{a model of `ReadablePropertyMap` with value type `geom_traits::Vector_3`}
-       \cgalParamDefault{If this parameter is omitted, normals are not written in the output stream.}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{geom_traits}
-       \cgalParamDescription{an instance of a geometric traits class}
-       \cgalParamType{a model of `Kernel`}
-       \cgalParamDefault{a \cgal Kernel deduced from the point type, using `CGAL::Kernel_traits`}
-     \cgalParamNEnd
-
-     \cgalParamNBegin{stream_precision}
-       \cgalParamDescription{a parameter used to set the precision (i.e. how many digits are generated) of the output stream}
-       \cgalParamType{int}
-       \cgalParamDefault{the precision of the stream `os`}
-       \cgalParamExtra{This parameter is only meaningful while using \ascii encoding.}
-     \cgalParamNEnd
-   \cgalNamedParamsEnd
-
-   \returns `true` if writing was successful, `false` otherwise.
-
-   \sa `write_PLY_with_properties()`
-*/
-template <typename PointRange, typename CGAL_NP_TEMPLATE_PARAMETERS>
+// documented in ../PLY.h
+template <typename PointRange, typename CGAL_NP_TEMPLATE_PARAMETERS_NO_DEFAULT>
 bool write_PLY(std::ostream& os,
                const PointRange& points,
-               const CGAL_NP_CLASS& np = parameters::default_values()
-#ifndef DOXYGEN_RUNNING
-               , std::enable_if_t<internal::is_Range<PointRange>::value>* = nullptr
-#endif
+               const CGAL_NP_CLASS& np
+               , std::enable_if_t<internal::is_Range<PointRange>::value>*
                )
 {
   using parameters::choose_parameter;
@@ -274,13 +157,11 @@ bool write_PLY(std::ostream& os,
 
    \sa `write_PLY_with_properties()`
 */
-template <typename PointRange, typename CGAL_NP_TEMPLATE_PARAMETERS>
+template <typename PointRange, typename CGAL_NP_TEMPLATE_PARAMETERS_NO_DEFAULT>
 bool write_PLY(const std::string& filename,
                const PointRange& points,
-               const CGAL_NP_CLASS& np = parameters::default_values()
-#ifndef DOXYGEN_RUNNING
-               , std::enable_if_t<internal::is_Range<PointRange>::value>* = nullptr
-#endif
+               const CGAL_NP_CLASS& np,
+               std::enable_if_t<internal::is_Range<PointRange>::value>*
                )
 {
   const bool binary = CGAL::parameters::choose_parameter(CGAL::parameters::get_parameter(np, internal_np::use_binary_mode), true);
