+++ Stream_support/include/CGAL/IO/XYZ.h
@@ -8,16 +8,17 @@
 //
 // Author(s)     : Simon Giraudot
 
-#ifndef CGAL_IO_LAS_H
-#define CGAL_IO_LAS_H
+#ifndef CGAL_IO_XYZ_H
+#define CGAL_IO_XYZ_H
 
-#include <CGAL/IO/LAS/read_las_points.h>
-#include <CGAL/IO/LAS/write_las_points.h>
 
 #include <fstream>
 #include <iostream>
 #include <vector>
 #include <type_traits>
+#include <CGAL/Named_function_parameters.h>
+#include <CGAL/boost/graph/named_params_helper.h>
+#include <CGAL/Kernel_traits.h>
 
 namespace CGAL {
 
@@ -28,111 +29,18 @@ namespace CGAL {
 namespace IO {
 
 /**
-   \ingroup PkgStreamSupportIoFuncsLAS
+   \ingroup PkgStreamSupportIoFuncsIOXyz
 
-   generates a %LAS property handler to read 3D points. Points are
-   constructed from the input the using 3 %LAS properties
-   `LAS_property::X`, `LAS_property::Y` and `LAS_property::Z`.
+   \brief reads points (positions + normals, if available), using the \ref IOStreamXYZ.
 
-   \tparam PointMap the property map used to store points.
-
-   \sa `read_LAS_with_properties()`
-   \sa \ref IOStreamLAS
-*/
-template <typename PointMap>
-std::tuple<PointMap,
-           typename Kernel_traits<typename PointMap::value_type>::Kernel::Construct_point_3,
-           LAS_property::X, LAS_property::Y, LAS_property::Z >
-make_las_point_reader(PointMap point_map);
-
-
-/**
-   \ingroup PkgStreamSupportIoFuncsLAS
-
-   \brief reads user-selected points properties from a .las or .laz stream.
-   Potential additional properties are ignored.
-
-   Properties are handled through a variadic list of property
-   handlers. A `PropertyHandler` can either be:
-
-   - A `std::pair<PropertyMap, LAS_property::Tag >` if the user wants to
-   read a %LAS property as a scalar value `LAS_property::Tag::type` (for
-   example, storing an `int` %LAS property into an `int` variable).
-
-   - A `std::tuple<PropertyMap, Constructor,
-   LAS_property::Tag...>` if the user wants to use one or several
-   %LAS properties to construct a complex object (for example,
-   storing 4 `unsigned short` %LAS properties into a %Color object
-   that can for example be a `std::array<unsigned short,
-   4>`). In that case, the second element of the tuple should be a
-   functor that constructs the value type of `PropertyMap` from N
-   objects of of type `LAS_property::Tag::type`.
-
-   The %LAS standard defines a fixed set of properties accessible
-   through the following tag classes:
-
-   - `LAS_property::X` with type `double`
-   - `LAS_property::Y` with type `double`
-   - `LAS_property::Z` with type `double`
-   - `LAS_property::Intensity` with type `unsigned short`
-   - `LAS_property::Return_number` with type `unsigned char`
-   - `LAS_property::Number_of_returns` with type `unsigned char`
-   - `LAS_property::Scan_direction_flag` with type `unsigned char`
-   - `LAS_property::Edge_of_flight_line` with type `unsigned char`
-   - `LAS_property::Classification` with type `unsigned char`
-   - `LAS_property::Synthetic_flag` with type `unsigned char`
-   - `LAS_property::Keypoint_flag` with type `unsigned char`
-   - `LAS_property::Withheld_flag` with type `unsigned char`
-   - `LAS_property::Scan_angle` with type `double`
-   - `LAS_property::User_data` with type `unsigned char`
-   - `LAS_property::Point_source_ID` with type `unsigned short`
-   - `LAS_property::Deleted_flag` with type `unsigned int`
-   - `LAS_property::GPS_time` with type `double`
-   - `LAS_property::R` with type `unsigned short`
-   - `LAS_property::G` with type `unsigned short`
-   - `LAS_property::B` with type `unsigned short`
-   - `LAS_property::I` with type `unsigned short`
-
-   \attention To read a binary file, the flag `std::ios::binary` must be set during the creation of the `ifstream`.
-
-   \tparam OutputIteratorValueType type of objects that can be put in `PointOutputIterator`.
-   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<PointOutputIterator>::%type`.
-   It can be omitted if the default is fine.
-   \tparam PointOutputIterator iterator over output points.
-   \tparam PropertyHandler handlers to recover properties.
-
-   \returns `true` if reading was successful, `false` otherwise.
-
-   \sa `make_las_point_reader()`
-
-   \sa \ref IOStreamLAS
-*/
-template <typename OutputIteratorValueType,
-          typename PointOutputIterator,
-          typename ... PropertyHandler>
-bool read_LAS_with_properties(std::istream& is,
-                              PointOutputIterator output,
-                              PropertyHandler&& ... properties);
-
-
-
-/**
-   \ingroup PkgStreamSupportIoFuncsLAS
-
-   \brief reads points (position only) using the \ref IOStreamLAS.
-
-   Potential additional properties are ignored.
-
-   \attention To read a binary file, the flag `std::ios::binary` must be set during the creation of the `ifstream`.
-
-   \tparam OutputIteratorValueType type of objects that can be put in `PointOutputIterator`.
-   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<PointOutputIterator>::%type`.
+   \tparam OutputIteratorValueType type of objects that can be put in `OutputIterator`.
+   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<OutputIterator>::%type`.
    It can be omitted when the default is fine.
-   \tparam PointOutputIterator iterator over output points.
+   \tparam OutputIterator iterator over output points.
    \tparam NamedParameters a sequence of \ref bgl_namedparameters "Named Parameters"
 
-   \param is input stream
-   \param output output iterator over points
+   \param is input stream.
+   \param output output iterator over points.
    \param np an optional sequence of \ref bgl_namedparameters "Named Parameters" among the ones listed below
 
    \cgalNamedParamsBegin
@@ -140,6 +48,12 @@ bool read_LAS_with_properties(std::istream& is,
        \cgalParamDescription{a property map associating points to the elements of the point range}
        \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Point_3`}
        \cgalParamDefault{`CGAL::Identity_property_map<geom_traits::Point_3>`}
+     \cgalParamNEnd
+
+     \cgalParamNBegin{normal_map}
+       \cgalParamDescription{a property map associating normals to the elements of the point range}
+       \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Vector_3`}
+       \cgalParamDefault{If this parameter is omitted, normals in the input stream are ignored.}
      \cgalParamNEnd
 
      \cgalParamNBegin{geom_traits}
@@ -151,40 +65,44 @@ bool read_LAS_with_properties(std::istream& is,
 
    \returns `true` if reading was successful, `false` otherwise.
 
-   \sa `read_LAS_with_properties()`
+   \sa \ref IOStreamXYZ
 */
 template <typename OutputIteratorValueType,
-          typename PointOutputIterator,
+          typename OutputIterator,
           typename CGAL_NP_TEMPLATE_PARAMETERS>
-bool read_LAS(std::istream& is,
-              PointOutputIterator output,
+bool read_XYZ(std::istream& is,
+              OutputIterator output,
               const CGAL_NP_CLASS& np = parameters::default_values());
 
 
 
 
 /**
-   \ingroup PkgStreamSupportIoFuncsLAS
+   \ingroup PkgStreamSupportIoFuncsIOXyz
 
-   \brief reads points (position only) using the \ref IOStreamLAS.
+   \brief reads points (positions + normals, if available), using the \ref IOStreamXYZ.
 
-   Potential additional properties are ignored.
-
-   \tparam OutputIteratorValueType type of objects that can be put in `PointOutputIterator`.
-   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<PointOutputIterator>::%type`.
+   \tparam OutputIteratorValueType type of objects that can be put in `OutputIterator`.
+   It must be a model of `DefaultConstructible` and defaults to `value_type_traits<OutputIterator>::%type`.
    It can be omitted when the default is fine.
-   \tparam PointOutputIterator iterator over output points.
+   \tparam OutputIterator iterator over output points.
    \tparam NamedParameters a sequence of \ref bgl_namedparameters "Named Parameters"
 
-   \param filename name of the input file
-   \param output output iterator over points
-   \param np an optional sequence of \ref bgl_namedparameters "Named Parameters" among the ones listed below
+   \param fname input file name.
+   \param output output iterator over points.
+   \param np optional sequence of \ref bgl_namedparameters "Named Parameters" among the ones listed below.
 
    \cgalNamedParamsBegin
      \cgalParamNBegin{point_map}
        \cgalParamDescription{a property map associating points to the elements of the point range}
        \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Point_3`}
        \cgalParamDefault{`CGAL::Identity_property_map<geom_traits::Point_3>`}
+     \cgalParamNEnd
+
+     \cgalParamNBegin{normal_map}
+       \cgalParamDescription{a property map associating normals to the elements of the point range}
+       \cgalParamType{a model of `WritablePropertyMap` with value type `geom_traits::Vector_3`}
+       \cgalParamDefault{If this parameter is omitted, normals in the input stream are ignored.}
      \cgalParamNEnd
 
      \cgalParamNBegin{geom_traits}
@@ -196,16 +114,19 @@ bool read_LAS(std::istream& is,
 
    \returns `true` if reading was successful, `false` otherwise.
 
-   \sa `read_LAS_with_properties()`
+   \sa \ref IOStreamXYZ
 */
 template <typename OutputIteratorValueType,
-          typename PointOutputIterator,
-          typename CGAL_NP_TEMPLATE_PARAMETERS>
-bool read_LAS(const std::string& filename,
-              PointOutputIterator output,
+          typename OutputIterator,
+           typename CGAL_NP_TEMPLATE_PARAMETERS>
+bool read_XYZ(const std::string& fname,
+              OutputIterator output,
               const CGAL_NP_CLASS& np = parameters::default_values());
 } // namespace IO
-
 } // namespace CGAL
 
-#endif // CGAL_IO_LAS_H
+
+#include <CGAL/IO/XYZ/read_xyz_points.h>
+#include <CGAL/IO/XYZ/write_xyz_points.h>
+
+#endif // CGAL_IO_XYZ_H
